+++ integration-tests/src/test/java/org/zowe/apiml/integration/authentication/schemes/X509SchemeTest.java
@@ -89,7 +89,7 @@ class X509SchemeTest implements TestWithStartedInstances {
                 .when()
                 .get(X509SchemeTest.URL)
                 .then()
-                .header("X-Zowe-Auth-Failure", is("ZWEAG164E Error occurred while validating X509 certificate. X509 certificate is missing the client certificate extended usage definition")).statusCode(200);
+                .header("X-Zowe-Auth-Failure", is("ZWEAG165E X509 certificate is missing the client certificate extended usage definition")).statusCode(200);
         }
 
         @Test
