+++ integration-tests/src/test/java/org/zowe/apiml/integration/penetration/JwtPenTest.java
@@ -62,7 +62,7 @@ import static org.zowe.apiml.util.SecurityUtils.getConfiguredSslConfig;
  *      - token signed with public key
  */
 @GeneralAuthenticationTest
-class JwtPenTest implements TestWithStartedInstances {
+public class JwtPenTest implements TestWithStartedInstances {
 
     private static final EnvironmentConfiguration ENV = ConfigReader.environmentConfiguration();
 
@@ -75,12 +75,9 @@ class JwtPenTest implements TestWithStartedInstances {
 
     private static final String BASE_PATH = "/api/v1/gateway";
     private static final String QUERY_ENDPOINT = "/auth/query";
-    private static final String PASSTICKETCLIENT_BASE_PATH = "/api/v1/dcpassticket";
-    private static final String PASSTICKET_TEST_ENDPOINT = "/passticketTest";
     private static final String PASSTICKET_ENDPOINT = "/auth/ticket";
 
     private static final String QUERY_ENDPOINT_URL = String.format("%s://%s:%d%s%s", SCHEME, HOST, PORT, BASE_PATH, QUERY_ENDPOINT);
-    private static final String PASSTICKET_ENDPOINT_URL = String.format("%s://%s:%d%s%s", SCHEME, HOST, PORT, PASSTICKETCLIENT_BASE_PATH, PASSTICKET_TEST_ENDPOINT);
     private static final String GATEWAY_ENDPOINT_URL = String.format("%s://%s:%d%s%s", SCHEME, HOST, PORT, BASE_PATH, PASSTICKET_ENDPOINT);
 
     private static String validToken;
@@ -171,7 +168,7 @@ class JwtPenTest implements TestWithStartedInstances {
         );
     }
 
-    private String getToken(String tokenType) {
+    public static String getToken(String tokenType) {
         switch (tokenType) {
             case "validJwt":
                 return getValidToken();
@@ -202,23 +199,6 @@ class JwtPenTest implements TestWithStartedInstances {
                     .header("Authorization", "Bearer " + token)
                 .when()
                     .get(QUERY_ENDPOINT_URL)
-                .then()
-                    .statusCode(is(status));
-                //@formatter:on
-
-            }
-
-            @ParameterizedTest(name = "call passticket service with {0} to receive response code {2}")
-            @MethodSource("org.zowe.apiml.integration.penetration.JwtPenTest#getTokens")
-            @TestsNotMeantForZowe
-            void whenCallPassTicketService(String tokenType, int status) {
-                String token = getToken(tokenType);
-
-                //@formatter:off
-                given()
-                    .header("Authorization", "Bearer " + token)
-                .when()
-                    .get(PASSTICKET_ENDPOINT_URL)
                 .then()
                     .statusCode(is(status));
                 //@formatter:on
